requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
distro==1.9.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.0
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
